pyproject.toml
[tool.poetry]
name = "madlibs"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
aiosqlite3 = "^0.3.0"
jishaku = "^1.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
